Solve few inconsistent typing issues via mypy.
This commit is the first trial to solve issues that appear via mypy.
From now on we should think about using mypy while developing.

Generally there is a big problem regarding the numbers package
(obviously this doesn't work with mypy). A change from numbers.Number
to numbers.Real may already help a bit, but for a long term goal
we should move to something like:

    Real = typing.Union[int, float, fractions.Fraction]