Make magic method __mul__ return ComplexEvent
All methods inherited from list and applied to ComplexEvent that
normally would return a list should return a ComplexEvent.
Before this commit this was true for the __add__ method, but not
for the __mul__ method.

Furthermore this commit adds two tests in basic_tests to make sure
the magic methods return the expected type.