Replace numbers type hints with mutwo type hints
Because of [this](https://github.com/python/mypy/issues/3186) mypy
error, mutwo will use a generic typing.Union variable for type hints
regarding numbers.Real for now. For isinstance checks mutwo will
still make use of the numbers module (because it is not possible to
typecheck with a typing.Union object).