Add again syntactic sugar in setter methods
When trying to solve mypy error messages with the last commits,
I changed the setter methods of mutwo.events.music.NoteLike so that
the setter argument types and getter argument types are equal.
This breaks a syntactic sugar of converting data to the expected
type. This issue is widely discussed
[here](https://github.com/python/mypy/issues/3004). It seems
that mypy will add this feature in the future. For now it
is unsupported by mypy, but the other comments in the mentioned
issue leaded to the decision to add the syntactic sugar in mutwo
again.