Replace EnvelopeEvent with expenvelope.Envelope
Before this commit mutwo sometimes used its own envelope representation
(a SequentialEvent filled with EnvelopeEvent objects). Starting from
this commit, mutwo will always rely on the expenvelope package from
Marc Evanstein. This has the advantage that in any submodule mutwo
can make use of the same Envelope representation (before it wouldn't
have been possible to use an Envelope in the parameters module
because events depend on parameters and this would have resulted
in a circular import). Furthermore this simplifies the conversion
routines.