Update calculation for midi velocity
Previously midi velocity has been calculated in reference to the Volumes
amplitude where an equal decrease of amplitude lead to an equivalent
decrease of midi velocity. This relationship is problematic, because
halving the amplitude is very different from making a sound half as
loud as before. The logarithmic scale of decibel fits better to the
human hearing and it seems to be closer to what midi synthesizer expect.