Add new prop to JI pitch and improve compute_lazy
for JI pitch:
add "cent_deviation_to_closest_western_pitch" property. This may help
when writing scores, so the cent deviation can be denoted.

for compute_lazy:
Try to use dill instead of pickle if possible
This improves the amount of objects which can be supported with the
"compute_lazy" decorator.