Remove one-argument-policy for convert method
Experience proved that the converters API gains a higher usability
if we give up the one-argument-policy for the convert method.
Initially we wanted to give users a simple and uniform experience
for converter classes (therefore always the same arguments for the
convert method). But when writing more and more converters and using
them in real-word-examples it turned out that in many cases we need
more than one argument for the convert method. The arguments for the
__init__ method should be mostly reserved for setting up the class in
a way that it can be used several times with this setup.