Remove abstract brackets module
The 'time_brackets' contains now all functionality which was previously
split to the parent classes in the 'brackets' module. Initially I had
an idea for alternative type of brackets (besides time_brackets), but
now I think the idea is so unlikely to be usable that the gain for
having everything in one module (simplifying development, testing, etc.)
is bigger than having the option for alternative type of brackets.