Remove abstract class Parameter
Initially I had the idea that an abstract base class Parameter may be
helpful for type hints. For now it is never used, therefore I remove
the class. Because it is only an empty class it would be easy to
reimplement it, if it turns out that it is actually useful.