Move music related params and conv to music ext
For making the core mutwo even slimmer and less arbitrary all music
related parameters and converters have been moved to mutwo.ext-music.
In this way the core mutwo only takes care of events and parameters
in terms of objects which are related to time (and which don't know
anything about music or sound).