Add __call__ method to core_converters.abc.Converter
The __call__ method  simply calls the `convert` method of the `Converter` class.
This method isn't indented to finally replace the `convert` method but to
allow converters to be passed as arguments when functions (callable
objects) are expected. In this way anonymous functions which are passed
to converters in order to extract information from events can be
standardised and shared between various packages.