Drop prime factor utilities
The current module depends on a no longer maintained python wrapper of
the "primesieves" package. In mutwo.core the prime functions aren't used
at all anyway. Any package which depends on prime utilities should use
sympy (which is indeed a bit overkill, but it will be maintained on a
long term and its installation process is well documented and it is
available on nix).