set/mutate_parameter:Fix for >1 refs of same event
Before this patch it was dangerous to call 'set_parameter' or
'mutate_parameter' on a 'mutwo.core_events.ComplexEvent' which contains
multiple references of the same event. It was dangerous, because
set_parameter would be called many times on the same event, which would
lead to unexpected results. For instance:

>>> simple_event = core_events.SimpleEvent(2)
>>> s = core_events.SequentialEvent([simple_event, simple_event])
>>> s.set_parameter('duration', lambda old_duration: old_duration * 2)
>>> s[0].duration
8

So instead of the expected return value '4' (which would be the
old_duration 2 multiplied by 2) we get 8, because 'set_parameter'
has been called twice since there are two references of the same event
'simple_event'.

After this patch the above code would return the expected 4.

This patch changes the API of inheriting from the abstract base class
'mutwo.core_events.abc.Event'. Instead of overriding the abstract
methods "set_parameter" and "mutate_parameter", users need to override
the methods "_set_parameter" and "_mutate_parameter". Because no outside
library currently depends on this ABC API detail (and 99.999% of all
time users will inherit from SimpleEvent or SequentialEvent or
SimultaneousEvent) this change is not considered to break the major
version of mutwo, but only to be a bug fix actually.

---

This patch fixes issue https://github.com/mutwo-org/mutwo.core/issues/23.
Additional context can be found there.