ComplexEvent getter/setter: Allow tag for index
Now it is possible to get and set child events of a 'ComplexEvent' (e.g.
items inside a SequentialEvent or SimultaenousEvent) via the tag of the
child event.

So it's for instance possible to do the following:

>>> s = core_events.SequentialEvent(
>>>     [core_events.TaggedSimpleEvent(1, tag='vl'), core_events.TaggedSimpleEvent(1, tag='va')]
>>> )
>>> vl_ev = s['vl']

This greatly simplifies working with tagged events, because it's
sufficient to remember the tag of an event and it's not necessary to
know its index.