SimultaneousEvent: Add concatenation methods
Before this patch it was very difficult to concatenate two
SimultaneousEvent along their time axis. It was only easily possible to
extend a SimultaneousEvent with another SimultaneousEvent by adding more
simultaneous layers via the '.extend' method.

With this patch it becomes easy to extend each child of a
SimultaneousEvent with the children of a similar second
SimultaneousEvent. The new methods 'concatenate_by_index' and
'concatenate_by_tag' find pairs of events in the original and in new
SimultaneousEvent according to their common tag or to their common
index and concatenates the pairs. The concatenation is mostly safe
to use, because it ensures that even slightly different SimultaneousEvent
(e.g. one of them has more children than the other) still get concatenated
in meaningful ways.