core_events/ComplexEvent: Fix deletion via tag (https://github.com/mutwo-org/mutwo.core/issues/25)
With https://github.com/mutwo-org/mutwo.core/commit/441e954c57216179a03063dabe103d17119bf4c4
we added the feature to get/set child events via their tag. But in
this commit we missed adding the functionality to remove events via
their tag. This leads to an inconsistent API which is fixed with this
commit.