TempoConverter: ↑ performance (drop useless computations)
Before this patch we made many calculations which actually where
completely useless, because if a tempo envelope doesn't change and has
tempo 60 it means that it doesn't affect the duration of the event. So
we don't need to apply the tempo envelope on it. This saves a lot of
time.