events/Envelope.split_at: Make more reliable
This patch makes envelopes 'split_at' method more stable and reliable by
not depending on implementation details of parent 'split_at' method.

Background
==========

If we split an envelope and want to be sure our split envelope is really
equal in meaning to our original envelope, we need to be sure that there
is a control point at the split position. Before this patch this was
verified because 'cut_off' made sure that there are control point at cut
positions and 'split_at' is internally implemented with 'cut_off'. But if
this cross-dependency doesn't exist anymore - which *could* happen -
'Envelope.split_at' would be broke.