split_at: Allow > 1 split times + ↑ perf for SeqEv
This patch changes 'split_at', so that it allows multiple split times.
This patch also changes the internal implementation of
'SequentialEvent.split_at' which makes it much more performant if
multiple split operations are used in the same step. Finally this also
improves the performance of 'SimultaneousEvent.sequentialize', because
it can make use of the new batch-processing-ability of '.split_at'.