SimEvent.concatenate_by_*: Don't copy other event
When concatenating a `SimultaneousEvent` to another `SimultaneousEvent`
this operation can be compared to the default `extend` operation of a
python list. When doing this operation the other sequence isn't deep
copied, e.g.:

    >>> a, b = [...], [x, y]
    >>> a.extend(b)
    >>> b[0].changeSomething()
    >>> a[2].hasSomethingChanged is True

Before this patch the 'concatenate_by_*' methods did deep copy the other
event. This patch fixes this bad habit and avoids deep copying the other
event. This change has multiple advantages:

1. Performance is much better, because deepcopy is a very expensive
   operation.

2. 'concatenate_by_*' behaves more pythonic since it's closer to how
   default python functions work.