events: Fix tempo_envelope concatenation
With https://github.com/mutwo-org/mutwo.core/commit/881b7e950518c9aa3342f446143401d2f688af1d we ensured that tempo envelopes of events
are also proceeded when two sequential events are concatenated.
https://github.com/mutwo-org/mutwo.core/issues/27 revealed there was a
mistake in the implementation: tempo envelopes aren't relative to the
events duration, but they are absolute. This also seems to be a more
intuitive approach for this, so we should keep it like this. This patch
fixes the package wide consistency by ensuring that when concatenating
two sequential or simultaneous event the tempo envelopes of both events
are correctly concatenated (e.g. if a tempo envelope is too long, it
will be truncated; if it is too short it will be extended so it fits
to the events duration).

This fixes https://github.com/mutwo-org/mutwo.core/issues/27