events += TempoEvent to be used in TempoEnvelope
'TempoEvent' is 'SimpleEvent' which additionally contains a
'TempoPoint' (as 'tempo_point'). It is supposed to be used as an
item in a 'TempoEnvelope'. It's better than using 'SimpleEvent' and
dynamically assigning a 'tempo_point' to this 'SimpleEvent', because
with a 'TempoEvent' we can support syntactic sugar in the setter of
the 'tempo_point' attribute, which allows us to ensure that we always
return a 'core_parameters.abc.TempoPoint' instance. This makes it safer
to use and with less surprises.