events: Make Envelope.extend_until more reliable
The method '.extend_until' makes an event longer if it isn't long
enough yet. For envelopes it's not the best idea to simply make the
last control point longer, because when then a new envelope is extended
the curve of the former envelope is broke. It's better if the last
control point has duration = 0 to ensure when adding new control points
the curve isn't changed before the new control points.