Refactor '...Event' to '...'
In https://github.com/mutwo-org/mutwo.core/issues/18 we refactored all
converters from '...Converter' to 'Format0ToFormat1'. This clarified the
functionality of the converter and reduced redundancy: that the class is
a converter is already clear from the module context (we have
'midi_CONVERTERS...', so all contained objects are converters). The
redundancy is also true for events. Consider the following code:

    from mutwo import core_events
    e = core_events.SequentialEvent([core_events.SimpleEvent()])

That 'SequentialEvent' and 'SimpleEvent' are events is already clear
from their module context: it already says 'core_EVENTS'.
So the 'Event' in their name is redundant. We can drop it and therefore
make code which uses mutwo less verbose. This makes naming also more
consistent with how events are named in 'mutwo.music': NoteLike (and not
'NoteLikeEvent').

Because this is a breaking change, until 'mutwo.core' 2.0.0 release,
this package still supports the old syntax with 'SequentialEvent'.