Drop 'mutate=False/True' option & decorator
Before this patch, many mutwo object methods used to have
a 'mutate' keyword. By default this keyword was set to 'True',
which meant that the internal state of the methods object was
changed. But the user had the option to set it to 'False', so
that a new copy of this object was created. In other words this
was a syntactic sugar for

    object.method(mutate=False) == object.copy().method()

The advantages of this syntactic sugar seem to be rather low,
as the above example shows that both ways of writing this
expression are more or less equally long. The 'mutate' keyword
syntax allowed methods to decide whether they internally use a
different implementation if copying is set or not, so that
performance-wise methods could improve. But in reality almost
all methods are internally implemented in a way which changes
the object and if 'mutate=False' almost all methods therefore
simply copied the object. The 'mutate' keyword was also very
hard to document, as it was often introduced with an implicit
decorator (this also made it difficult for type checker to
understand what happens).

Summarized the 'mutate' syntax didn't help with either internal
optimization nor did it help the user with a better API. The
'copy' syntax is much more explicit and easier to understand.
So in order so simplify the code base for a better long-term
maintenance, let's drop the 'mutate' keyword and the related
decorator 'add_copy_option'.