utilities += freezables
Python objects are either mutable or immutable. Generally
Python has an approach of favoring mutability (no real
privacy exists, 'we are all adults here') & 'mutwo' follows
this general approach by providing mostly mutable objects.

In Python no general functionality exists to transform mutable
objects into immutable objects. This is different in other
languages: Ruby objects know the 'freeze' method [1], also
Javascript object can be frozen [2]. In Java a similar concept
exists with the 'final' keyword [3].

This patch aims to add a similar concept to Python via the
'Freezable' class.

Why do we need this concept?

Mutability of objects can help to improve the performance of some
operations: if we only want to change one element of a very long
array, it's more efficient to only change this one element in-place
(mutable) instead of copying the whole array & changing only one
element in this copy-process (immutable).

On the other hand, some situations are inefficient if objects are
mutable: if an object has various dynamic (= mutable) properties
that are used to calculate a specific value, this value always
needs to be re-calculated even if its dependencies - the dynamic
properties - never change. In such a case - when we know we never
need to change our dynamic properties - it's much more efficient
to only call the computation that results in the searched value
once & cache this value then for further usages [4], [5].

We see that in different situations mutability or immutability
can be more efficient. In order to support both situations, this
patch adds the new 'mutwo.core_utilities.Freezable' class.
Any class can inherit from this base class in order to switch
between mutable/immutable states. By default classes are mutable,
but they become immutable once the 'freeze` method has been called.

[1] https://ruby-doc.org/core-2.5.0/Object.html#method-i-freeze
[2] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/freeze
[3] https://en.wikipedia.org/wiki/Final_(Java)
[4] In [3] the text goes:

    "Like full immutability, the use of final
    variables has great advantages, especially in optimization. For
    instance, Sphere will probably have a function returning its volume;
    knowing that its radius is constant allows us to memoize the computed
    volume. If we have relatively few Spheres and we need their volumes very
    often, the performance gain might be substantial. Making the radius of a
    Sphere final informs developers and compilers that this sort of
    optimization is possible in all code that uses Spheres."

[5] In 'mutwo' we encountered such situation in https://github.com/mutwo-org/mutwo.core/commit/05711a72069ce81fa29254b005e9b79a141d4d90.