events/Envelope: Drop deprecated Envelope.from_points
The 'from_points' classmethod was initally added to simplify the
migration from the 'expenvelope' package to mutwos builtin envelope
classes. This migration is already more than a year ago [1], and the
docstring of the 'from_points' method mentioned for a long time its
deprecation; no one should rely on this anymore. Therefore let's cleanup
the source code and drop this method.

[1] https://github.com/mutwo-org/mutwo.core/commit/3751cd694e746e5e3d33cccf895aedf9f538645d