ROADMAP: Drop refactor DirectParameter => Parameter
It looks tempting to refactor all 'DirectX' to 'X' (e.g.
'DirectDuration' to 'Duration'), as in mutwo the most common import
pattern is:

>>> from mutwo import x_parameters
>>> x_parameters.DirectX
>>> x_parameters.abc.X

So there is mostly no conflict between 'X' and 'DirectX'.
But this is only true when staying within this mutwo convention.
A user is free to do:

>>> from mutwo.x_parameters import DirectX
>>> from mutwo.x_parameters.abc import X

In this case the refactor from 'DirectX' to 'X' would result in a
namespace conflict between the abstract base class 'X' and the class
'X'.

So although this simplification looks elegant, it's actually not very
pythonic as it forces people to use a mutwo specific import policy. It
seems to create less conflict by just leaving everything as it is.