Fix flaky tests
https://github.com/mutwo-org/mutwo.core/commit/20f4256cec09c92eff3b1c3cce40c931fa2a4faa introduced unit test style
performance tests to automatically detect performance regressions.
Unfortunately theses tests are very flaky, as the exact timing result
depends on a lot of environmental factors that aren't controlled within
the test. Therefore this patch moves the performance tests from the
global tests directory to a dedicated performance test directory, that
can be executed manually. This performance test directory is merged with
the old manual performance tests.