events/Envelope += time_range_to_point_tuple
This new function is useful if one wants to do maths with control points
over a given range - e.g. for manually calculating the integral in a
given time range. Because we want to do this in the next commit, this
patch prepares the Envelope class to already support finding all control
points.