WIP events/Envelope/configurations: Use default functions instead of names
This patch changes envelopes by using global default functions in the
conconfigurations submodule, instead of configurable names for event attributes.

---

Because this patch breaks backwards compatibility (e.g. removes the
public API, this should be moved to the v.2.0.0 branch!)

This is WIP because we also need to adapt TempoEnvelope (and maybe
RelativeEnvelope?).