Drop core_generators
Nowadays it looks very arbitrary why this specific algorithm is part of
'mutwo.core'. It's not needed by any other part of the project, but
includes the heavy dependency 'numpy' (also it may be usable without
numpy). But anyway, it doesn't belong to 'mutwo.core', but should rather
move to 'mutwo.common' or to its own repository. 'mutwo.core' shouldn't
be responsible for hosting fancy algorithmic-arts tools, but only the
bare-bones necessities to describe events, parameters and converters.