events: Drop TaggedXEvent => Add tag to abc.Event
This patch drops the special 'TaggedSimpleEvent',
'TaggedSequentialEvent' and 'TaggedSimultaneousEvent' classes and the
related decorator 'add_tag_to_class'. The patch also adds the 'tag'
attribute to 'abc.Event', so that all events now have a tag attribute.
If this 'tag' attribute is unset, it is 'None' and doesn't bother the
events. If one wants to set it, we no longer need special classes for
this. This patch therefore simplifies the codebase. It also becomes
easier to switch between a tagged event and a not-tagged event.