Simplify __repr__ and add concise __str__
The string representation of mutwo events is very verbose: often it is
so verbose that it's very difficult to debug from the printed data.
This patch improves the situation by making the present __repr__
slightly more concise and by adding additional __str__ methods to mutwo
objects, so that with 'print(X)' we can have a more concise
pretty-printed representation.