utilities/exceptions: Drop 'UndefinedReferenceWarning'
This warning had the function to inform the user if she/he/* used a
tempo point without a reference. Then this reference was automatically
set to 1. Unfortunately this warning isn't very useful. Very often such
tempo points are used and then the warnings just pollute the logging.
They may also look frightening to new users, while they can simply be
ignored. So it's better to drop this warning.