Improve readability of code: make code more concise
This patch doesn't change any semantics in the source code, but it
nevertheless makes the source code much smaller by using more concise
variable names and logics. In the early days of 'mutwo' there was an
abbreviation ban, because of bad experiences we've made from other
libraries where it was difficult to remember and understand the API
due to the usage of (sometimes even not coherent usage of)
abbreviations. Due to this most of the old code is very verbose.
Unfortunately this verbosity can also lead to text pollution and make a
code even less unreadable. This is true for code inside functions: here
the context is often clear, we only use variables which are never part
of the public API and it becomes better to read if we use shorter names.
Therefore this patch therefore also relax mutwo abbreviations ban to the
ban of abbreviations in the public API.