Generalize & simplify syntactic sugar parameter parser
Sometimes it's very convenient to use syntactic sugar, e.g. for writing
a chord it's nicer to type

    >>> music_events.NoteLike('c4 e5', '7/8', 'ff')

than to write

    >>> music_events.NoteLike(
    ...     [music_parameters.WesternPitch('c4'), music_parameters.WesternPitch('e5')],
    ...     core_parameters.RatioDuration(fractions.Fraction(7, 8)),
    ...     music_parameters.WesternVolume('ff')
    ... )

The conversion from strings to mutwo objects is done by public
functions. Before this patch these functions usually resided in the
'configurations' submodule of an event module. For instance the duration
parser was:

    >>> core_events.configurations.UNKNOWN_OBJECT_TO_DURATION

It's position in the events configurations module should encourage
users to override the definition with their own function.  It also resided
there in order to make the parsing explicit and not hidden from the user,
following the Python Zen 'Explicit is better than implicit' (see also
https://github.com/mutwo-org/mutwo.music/commit/bf47a452d8553ef001c5192393d697d0c9536dd2 for more context).

This patch makes a change from

    >>> core_events.configurations.UNKNOWN_OBJECT_TO_DURATION(?)
    >>> core_events.configurations.UNKNOWN_OBJECT_TO_TEMPO_POINT(?)

to

    >>> core_events.abc.Duration.from_any(?)
    >>> core_events.abc.TempoPoint.from_any(?)

By moving the functions to the parameter classes their position becomes
easier and more intuitive to understand. Furthermore they are less
verbose and faster to type / better to read. It also doesn't make a
technical difference for a user to monkey patch
'core_events.configurations.UNKNOWN_OBJECT_TO_DURATION' or to monkey
patch 'core_events.abc.Duration.from_any', so it's still possible to
configure the parsing (or to even completely disable it if a user
doesn't want it).

This patch also adds

    core_events.abc.Duration.Type
    core_events.abc.TempoPoint.Type
    ...

These are useful constants for type hints: these type aliases
'Duration.Type', 'TempoPoint.Type', ... should include all types that are
supported by the syntactic-sugar-parser. In this way it's easy to
document whether a function expects the mutwo object, or can also handle
other types and transforms them to mutwo objects. In this way it's
easier to make mypy happier. Perhaps this was also the initial reason
why 'core_constants.DurationType' was added (that has been removed in
https://github.com/mutwo-org/mutwo.core/commit/0f4f79257df520ab7d1b2a7809edaab600c2dc59).

Finally this patch drops

    - core_converters.UnknownObjectToObject

This converter was introduced in https://github.com/mutwo-org/mutwo.core/commit/d9a861c54e56af4c6886293c61f703c8a8d344a9 in
summer 2022. At this time 'mutwo' still used Python 3.9 which didn't
support the 'match..case' - syntax yet. With this new syntax it's very
easy to express the parsing of unknown objects into a specific type and
the converter 'core_converters.UnknownObjectToObject' is no longer
needed.