parameters/TempoPoint: Refactor 'tempo_in_beats_per_minute' to 'tempo'
It's an inconsistency of the 'TempoPoint' class to explicitly name its
attribute with the unit of its attribute. In other words we don't have

    Duration.duration_in_beats
    Pitch.frequency_in_hertz
    PitchInterval.interval_in_cents

but

    Duration.duration
    Pitch.frequency
    PitchInterval.interval

Therefore it also shouldn't be

    Tempo.tempo_in_beats_per_minute

but

    Tempo.tempo

This patch therefore helps unifies the naming of mutwo which makes it
easier for anyone new coming to the library.