parameters/TempoPoint: Neutralize + add WesternTempoPoint
ABSTRACT
========

This patch neutralizes the 'TempoPoint' parameter from western notation
influences by dropping specifications which are only needed in western
music notation. Furthermore it adds a new class 'WesternTempoPoint' with
which these specifications are added again.

DETAILS
=======

In western music notation it's common to indicate a tempo as a range
from the minimal (slowest) accepted tempo to the maximum (fastest)
accepted tempo. It's furthermore common to use a reference (1/4 of a
beat, 1/2 of a beat, ...) to which the tempo refers to. Finally western
notation often uses a textual indicator to describe the tempo (e.g.
'Adagio', 'Andante', etc.). All of these customs became part of the
general 'TempoPoint' implementation in https://github.com/mutwo-org/mutwo.core/commit/aa11536ecb68f8e29053ead03819ad789acadd1d
and https://github.com/mutwo-org/mutwo.core/commit/7d89b7da4b3d7dc71480d6999da0a032f9969cf8 and https://github.com/mutwo-org/mutwo.core/commit/f506c8c3be6506c2df128c1e14fa5cbefdb499f0.
The presence of these western notation customs in the abstact base class
of 'TempoPoint' was as mistake, as mutwo generally aims at neutral base
classes that aren't too influenced by specific traditions. Therefore
their presence needed to be removed.

Now 'TempoPoint' and 'DirectTempoPoint' are much simpler as they only
have a 'tempo' attribute that's in beats per minute (BPM). This also
helped simplifying some bits in the tempo converter class.

Without the 'reference' attribute, the 'tempo' attribute now always
specifies the tempo for 1/1 (1). So in a 'tempo' of 60 a 'duration' of
1 equals 1 second. This is different to what is usually thought of in
western music, where we think that BPM=60 means that one quarter note
(the default 'reference') equals a second. Therefore the default
'reference' of 'WesternTempoPoint' is set to '1/4' and therefore

    WesternTempoPoint(60) != DirectTempoPoint(60)

This may look confusing, but in case people want to work with western
models, it makes sense.

MORE DETAILS: CONTINGENCY
=========================

What's particularly striking about the 'TempoPoint' /
'WesternTempoPoint' case, is the presence of contingency.
Most parts of mutwo conceptually avoid contingency by favoring
explicit values: a pitch has a discreet frequency, a concurrence
hosts a discreet number of children events. This is different to
the concept of the 'tempo_range' attribute of 'WesternTempoPoint':
in this case the explicit tempo is not discreet, but it's an
infinite set of possible tempos. This raises the question if mutwo
shouldn't generally support contingency and uncertainty e.g. if it
would make sense that all pitch objects have a 'frequency contingency',
as a space of possible values for frequency (even if this space only
contains one possibility). In western music tempo is often described
as a contingency, maybe there is other music where pitch is described
as a contingency instead of a discreet value? In this case it would
be suitable if mutwo could generally represent these interpretations,
as this is the prime concern of the library.

Unfortunately contingency introduces complexity: there isn't one type
of contingency (as ranges), but infinite (a set to pick from, numbers
with uncertain parts, ...). Furthermore how should one calculate with
the different contingencies? Finally mutwo also aims at simplicity and
maintainability and the introduction of contingency into all abstract
base classes may be very complex. So it seems to be a better solution to
introduce contingency explicitly wherever it's needed to comprehend a
given interpretation of reality. It may be a good idea for the future to
introduce a general 'Contingency' type for the contingency itself: this
should be a class which can be resolved to a discreet value. In this way
the resolution and its API could be standardized. In the current
implementation the resolution of the 'WesternTempoPoint.tempo_range' is
inside the 'WesternTempoPoint.tempo' property and uses the strategy to
always use the minimal value.