parameters: Refactor *TempoPoint to *Tempo
Perhaps the reason why it was called 'TempoPoint' instead of 'Tempo' was
that 'Tempo' is a part of a 'TempoEnvelope' and therefore only a point
in the wider tempo trajectory. On the other hand the same could be said
of a pitch: also pitches can be continuously changing ('glissando'), but
we don't call them 'PitchPoints'. The difference is perhaps that the
pitch attribute has an optionally activated '.envelope' attribute, while
'Tempo' hasn't. But we should rather fix this and make this uniform
instead of having a specific *Point parameter.